refactor: resturcture rust folder stucture
