+++ src-tauri/Cargo.toml
@@ -22,7 +22,7 @@ serde_json = "1"
 
 tauri-plugin-system-info = { git = "https://github.com/HuakunShen/tauri-plugin-system-info", branch = "v2" } # use v2 branch for Tauri v2 plugin
 
-async-openai = { version = "0.23.3" }
+async-openai = { version = "0.28.0" }
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
 tokio = { version = "1", features = ["full"] }
 http = "0.2.12"
