+++ src-tauri/src/common/mod.rs
@@ -1,0 +1,2 @@
+
+pub mod http_client;
